Implement grpcClient and correct grpcService and modify proto file
